+++ CHANGES.md
@@ -21,7 +21,7 @@ Changes:
 -
 
 Fixes:
-- 
+-
 
 ## Current