+++ CHANGES.md
@@ -17,10 +17,10 @@ The following changes are not yet released, but are code complete:
 * Add scrapers for Texas Supreme Court, Court of Criminal Appeals, and Courts of Appeal
 
 Features:
--
+- Add new `handle_email` method to `SCOTUSEmail` to confirm subscriptions or fetch full dockets depending on email type #1712
 
 Changes:
--
+- BREAKING CHANGE: Output format of `SCOTUSEmail` changed
 
 Fixes:
 -